moving app into folder in containers so to be able to experiment with dynamic loading
